Suricata
Migrated the checkWriteNoRead function and it's test to the detection namespace for other engines.

Added the logic to suricata.
